refactor(front): split about form
